frontend: fix ts errors
